requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
python-dotenv

# Для локального
torch
transformers
pillow